Refactor materialize to be static method on Store
